+++ changelog.rst
@@ -7,6 +7,7 @@ For a complete changelog, see:
 * Diff block attribute `ignoreLine` changed to `ignore`. Being able to ingnore just a portion of the line. The script file did need to adapt.
 * Run block attribute 'exitcode' no longer accepts value `None`. It is now written `null` as this is the standart
 * Run block now suppors attribute `shell`. It allwos to run a command in a selected shell. Supported shells are: `cmd`, `sh` and `powershell`
+* New command line argument added `scriptName`. It allows to change the default script name.
 
 1.1.3
 -----